Moved pg-index-health-test-starter to separate module
