+++ .gitignore
@@ -1,3 +1,4 @@
 /.idea/
 /.gradle/
 /build/
+/pg-index-health-test-starter/build/
