+++ build.gradle.kts
@@ -2,166 +2,163 @@ import com.github.benmanes.gradle.versions.updates.DependencyUpdatesTask
 import com.github.spotbugs.snom.Confidence
 import com.github.spotbugs.snom.Effort
 import com.github.spotbugs.snom.SpotBugsTask
-import info.solidsoft.gradle.pitest.PitestTask
 import net.ltgt.gradle.errorprone.errorprone
 import org.sonarqube.gradle.SonarTask
 
 plugins {
-    id("java-library")
-    id("maven-publish")
-    id("signing")
+    id("java")
     id("jacoco")
     id("com.github.spotbugs") version "5.0.14"
     id("checkstyle")
     id("pmd")
     id("org.sonarqube") version "4.3.0.3225"
-    id("info.solidsoft.pitest") version "1.9.11"
     id("net.ltgt.errorprone") version "3.1.0"
     id("com.github.ben-manes.versions") version "0.47.0"
 }
 
-group = "io.github.mfvanek"
-version = "0.9.4-SNAPSHOT"
+description = "pg-index-health-test-starter build"
 
-repositories {
-    mavenLocal()
-    mavenCentral()
-}
+allprojects {
+    group = "io.github.mfvanek"
+    version = "0.9.4-SNAPSHOT"
 
-dependencies {
-    api("io.github.mfvanek:pg-index-health:0.9.3")
-    implementation(libs.spring.boot.starter)
-    annotationProcessor(libs.spring.boot.autoconfigureProcessor)
-    annotationProcessor(libs.spring.boot.configurationProcessor)
-
-    testImplementation(libs.spring.boot.starterTest)
-    testImplementation(platform("org.junit:junit-bom:5.10.0"))
-    testImplementation("org.junit.jupiter:junit-jupiter-api")
-    testImplementation("org.apache.commons:commons-text:1.10.0")
-    testImplementation("com.google.code.findbugs:jsr305:3.0.2")
-    testImplementation("org.assertj:assertj-core:3.24.2")
-    testImplementation("org.postgresql:postgresql:42.6.0")
-
-    testRuntimeOnly("org.junit.jupiter:junit-jupiter-engine")
-    testRuntimeOnly("org.junit.platform:junit-platform-launcher") {
-        because("required for pitest")
-    }
-
-    pitest(libs.pitest.dashboard.reporter)
-    checkstyle("com.thomasjensen.checkstyle.addons:checkstyle-addons:7.0.1")
-    errorprone("com.google.errorprone:error_prone_core:2.20.0")
-}
-
-java {
-    sourceCompatibility = JavaVersion.VERSION_11
-    targetCompatibility = JavaVersion.VERSION_11
-    withJavadocJar()
-    withSourcesJar()
-}
-tasks.withType<JavaCompile>().configureEach {
-    options.errorprone {
-        disableWarningsInGeneratedCode.set(true)
+    repositories {
+        mavenLocal()
+        mavenCentral()
     }
 }
 
-jacoco {
-    toolVersion = libs.versions.jacoco.get()
-}
+subprojects {
+    apply(plugin = "java")
+    apply(plugin = "org.sonarqube")
+    apply(plugin = "checkstyle")
+    apply(plugin = "pmd")
+    apply(plugin = "com.github.spotbugs")
+    apply(plugin = "jacoco")
+    apply(plugin = "net.ltgt.errorprone")
 
-tasks {
-    test {
-        useJUnitPlatform()
-        dependsOn(checkstyleMain, checkstyleTest, pmdMain, pmdTest, spotbugsMain, spotbugsTest)
-        finalizedBy(jacocoTestReport, jacocoTestCoverageVerification)
+    dependencies {
+        testImplementation(platform("org.junit:junit-bom:5.10.0"))
+
+        checkstyle("com.thomasjensen.checkstyle.addons:checkstyle-addons:7.0.1")
+        errorprone("com.google.errorprone:error_prone_core:2.20.0")
     }
 
-    jar {
-        manifest {
-            attributes["Implementation-Title"] = project.name
-            attributes["Implementation-Version"] = project.version
-        }
-        isPreserveFileTimestamps = false
-        isReproducibleFileOrder = true
+    java {
+        sourceCompatibility = JavaVersion.VERSION_11
+        targetCompatibility = JavaVersion.VERSION_11
+        withJavadocJar()
+        withSourcesJar()
     }
-
-    javadoc {
-        if (JavaVersion.current().isJava9Compatible) {
-            (options as StandardJavadocDocletOptions).addBooleanOption("html5", true)
+    tasks.withType<JavaCompile>().configureEach {
+        options.errorprone {
+            disableWarningsInGeneratedCode.set(true)
         }
     }
 
-    jacocoTestReport {
-        dependsOn(test)
+    tasks {
+        test {
+            useJUnitPlatform()
+            dependsOn(checkstyleMain, checkstyleTest, pmdMain, pmdTest, spotbugsMain, spotbugsTest)
+            finalizedBy(jacocoTestReport, jacocoTestCoverageVerification)
+        }
+
+        jar {
+            manifest {
+                attributes["Implementation-Title"] = project.name
+                attributes["Implementation-Version"] = project.version
+            }
+            isPreserveFileTimestamps = false
+            isReproducibleFileOrder = true
+        }
+
+        javadoc {
+            if (JavaVersion.current().isJava9Compatible) {
+                (options as StandardJavadocDocletOptions).addBooleanOption("html5", true)
+            }
+        }
+
+        jacocoTestReport {
+            dependsOn(test)
+            reports {
+                xml.required.set(true)
+                html.required.set(true)
+            }
+        }
+
+        jacocoTestCoverageVerification {
+            dependsOn(jacocoTestReport)
+            violationRules {
+                rule {
+                    limit {
+                        counter = "CLASS"
+                        value = "MISSEDCOUNT"
+                        maximum = "0.0".toBigDecimal()
+                    }
+                }
+                rule {
+                    limit {
+                        counter = "METHOD"
+                        value = "MISSEDCOUNT"
+                        maximum = "0.0".toBigDecimal()
+                    }
+                }
+                rule {
+                    limit {
+                        counter = "LINE"
+                        value = "MISSEDCOUNT"
+                        maximum = "0.0".toBigDecimal()
+                    }
+                }
+                rule {
+                    limit {
+                        counter = "INSTRUCTION"
+                        value = "COVEREDRATIO"
+                        minimum = "1.0".toBigDecimal()
+                    }
+                }
+            }
+        }
+
+        check {
+            dependsOn(jacocoTestCoverageVerification)
+        }
+
+        withType<SonarTask>().configureEach {
+            dependsOn(test, jacocoTestReport)
+        }
+    }
+
+    jacoco {
+        toolVersion = rootProject.libs.versions.jacoco.get()
+    }
+
+    checkstyle {
+        toolVersion = rootProject.libs.versions.checkstyle.get()
+        configFile = file("../config/checkstyle/checkstyle.xml")
+        isIgnoreFailures = false
+        maxWarnings = 0
+        maxErrors = 0
+    }
+
+    pmd {
+        toolVersion = rootProject.libs.versions.pmd.get()
+        isConsoleOutput = true
+        ruleSetFiles = files("../config/pmd/pmd.xml")
+        ruleSets = listOf()
+    }
+
+    spotbugs {
+        showProgress.set(true)
+        effort.set(Effort.MAX)
+        reportLevel.set(Confidence.LOW)
+        excludeFilter.set(file("../config/spotbugs/exclude.xml"))
+    }
+    tasks.withType<SpotBugsTask>().configureEach {
         reports {
-            xml.required.set(true)
-            html.required.set(true)
+            create("xml") { enabled = true }
+            create("html") { enabled = true }
         }
-    }
-
-    jacocoTestCoverageVerification {
-        dependsOn(jacocoTestReport)
-        violationRules {
-            rule {
-                limit {
-                    counter = "CLASS"
-                    value = "MISSEDCOUNT"
-                    maximum = "0.0".toBigDecimal()
-                }
-            }
-            rule {
-                limit {
-                    counter = "METHOD"
-                    value = "MISSEDCOUNT"
-                    maximum = "0.0".toBigDecimal()
-                }
-            }
-            rule {
-                limit {
-                    counter = "LINE"
-                    value = "MISSEDCOUNT"
-                    maximum = "0.0".toBigDecimal()
-                }
-            }
-            rule {
-                limit {
-                    counter = "INSTRUCTION"
-                    value = "COVEREDRATIO"
-                    minimum = "1.0".toBigDecimal()
-                }
-            }
-        }
-    }
-
-    check {
-        dependsOn(jacocoTestCoverageVerification)
-    }
-}
-
-checkstyle {
-    toolVersion = libs.versions.checkstyle.get()
-    configFile = file("config/checkstyle/checkstyle.xml")
-    isIgnoreFailures = false
-    maxWarnings = 0
-    maxErrors = 0
-}
-
-pmd {
-    toolVersion = libs.versions.pmd.get()
-    isConsoleOutput = true
-    ruleSetFiles = files("config/pmd/pmd.xml")
-    ruleSets = listOf()
-}
-
-spotbugs {
-    showProgress.set(true)
-    effort.set(Effort.MAX)
-    reportLevel.set(Confidence.LOW)
-    excludeFilter.set(file("config/spotbugs/exclude.xml"))
-}
-tasks.withType<SpotBugsTask>().configureEach {
-    reports {
-        create("xml") { enabled = true }
-        create("html") { enabled = true }
     }
 }
 
@@ -170,93 +167,6 @@ sonar {
         property("sonar.projectKey", "mfvanek_pg-index-health-test-starter")
         property("sonar.organization", "mfvanek")
         property("sonar.host.url", "https://sonarcloud.io")
-    }
-}
-tasks.withType<SonarTask>().configureEach {
-    dependsOn(tasks.test, tasks.jacocoTestReport)
-}
-
-pitest {
-    verbosity.set("DEFAULT")
-    junit5PluginVersion.set(libs.versions.pitest.junit5Plugin.get())
-    pitestVersion.set(libs.versions.pitest.core.get())
-    threads.set(4)
-    if (System.getenv("STRYKER_DASHBOARD_API_KEY") != null) {
-        outputFormats.set(setOf("stryker-dashboard"))
-    } else {
-        outputFormats.set(setOf("HTML"))
-    }
-    timestampedReports.set(false)
-    mutationThreshold.set(100)
-}
-tasks.withType<PitestTask>().configureEach {
-    mustRunAfter(tasks.test)
-}
-tasks.build {
-    dependsOn("pitest")
-}
-
-publishing {
-    publications {
-        create<MavenPublication>("mavenJava") {
-            from(components["java"])
-            versionMapping {
-                usage("java-api") {
-                    fromResolutionOf("runtimeClasspath")
-                }
-                usage("java-runtime") {
-                    fromResolutionResult()
-                }
-            }
-
-            pom {
-                name.set(project.name)
-                description.set("Spring Boot Starter for pg-index-health library")
-                url.set("https://github.com/mfvanek/pg-index-health-test-starter")
-                licenses {
-                    license {
-                        name.set("Apache License Version 2.0")
-                        url.set("https://www.apache.org/licenses/LICENSE-2.0")
-                    }
-                }
-
-                developers {
-                    developer {
-                        id.set("mfvanek")
-                        name.set("Ivan Vakhrushev")
-                        email.set("mfvanek@gmail.com")
-                    }
-                }
-
-                scm {
-                    connection.set("scm:git:https://github.com/mfvanek/pg-index-health-test-starter.git")
-                    developerConnection.set("scm:git@github.com:mfvanek/pg-index-health-test-starter.git")
-                    url.set("https://github.com/mfvanek/pg-index-health-test-starter")
-                }
-            }
-        }
-    }
-
-    repositories {
-        maven {
-            val releasesRepoUrl = uri("https://oss.sonatype.org/service/local/staging/deploy/maven2")
-            val snapshotsRepoUrl = uri("https://oss.sonatype.org/content/repositories/snapshots")
-            url = if (version.toString().endsWith("SNAPSHOT")) snapshotsRepoUrl else releasesRepoUrl
-
-            val sonatypeUsername: String by project
-            val sonatypePassword: String by project
-            credentials {
-                username = sonatypeUsername
-                password = sonatypePassword
-            }
-        }
-    }
-}
-
-signing {
-    if (!version.toString().endsWith("SNAPSHOT")) {
-        useGpgCmd()
-        sign(publishing.publications["mavenJava"])
     }
 }
